requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
requests==2.32.3
matplotlib==3.8.4



# ML
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.2


# RAG
langchain==0.3.4
langchain-community==0.3.3
langchain-openai==0.2.2
langchain-text-splitters==0.3.1
chromadb==0.5.5
rank-bm25==0.2.2
openai==1.45.0


# UI
streamlit==1.38.0


# Dev
python-dotenv==1.0.1